Parallelism
The whole simulation process has been parallelised for several ptocesses and trying to acquire all the computing power of a machine to make a process of calculation as fast as its possible. Thus the Classes.Simulation.simulate() function has been moved to "Unused" and replaced with Classes.Simulate.part_sim(start, stop, write) function that processes the chunk of a timeline specified with start and stop parameters. additionaly to that the main file has been changed according to multiprocessing library requirements.
Moreover the Classes.Simulation.export() function has been changed. Now it can take an argument with the np.array to be writen.